requirements.txt
anyio==4.12.1
av==16.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
ctranslate2==4.7.1
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.24.1
packaging==26.0
protobuf==6.33.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
sarvamai==0.1.25
shellingham==1.5.4
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
